requirements.txt
appdirs==1.4.3
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-Compress==1.4.0
Flask-Humanize==0.3.0
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
humanize==0.5.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
pbr==3.1.1
psutil==5.3.1
pyparsing==2.2.0
python-utils==2.2.0
six==1.11.0
SQLAlchemy==1.1.14
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
Werkzeug==0.12.2
WTForms==2.1
